Separated into result and decorator modules
